+++ src/interactors/buildTable/removeEmpty.js
@@ -1,0 +1,4 @@
+const removeEmpty = (entries, mainStats) => entries
+  .filter((entry) => mainStats.some((stat) => !!entry.stats[stat]));
+
+module.exports = removeEmpty;